+++ CHANGELOG.md
@@ -172,3 +172,7 @@ v0.21 + 1
 
 * The local transport now generates textual progress output like
   git-upload-pack does ("counting objects").
+
+* The git_submodule_update function was renamed to
+  git_submodule_update_strategy. git_submodule_update is now used to
+  provide functionalty similar to "git submodule update".